RestoredSource/requirements.txt
requests==2.31.0
pywebview==5.1
pycryptodome==3.20.0
